// Qanat.Web/src/app/shared/generated/model/update-water-account-parcels-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class UpdateWaterAccountParcelsDto { 
    ParcelIDs?: Array<number>;
    EffectiveYear: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface UpdateWaterAccountParcelsDtoForm { 
    ParcelIDs?: FormControl<Array<number>>;
    EffectiveYear: FormControl<number>;
}

export class UpdateWaterAccountParcelsDtoFormControls { 
    public static ParcelIDs = (value: FormControlState<Array<number>> | Array<number> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<number>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static EffectiveYear = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
}